Optimise decorators
Not all decorators need to be applied in list view (and involve potentially expensive data operations when applied to a list), so split the decorators into separate modules to be able to configure which are applied to lists, and which are only for single-task requests.
