+++ lib/api.js
@@ -1,10 +1,11 @@
 const Api = require('@asl/service/api');
 const Taskflow = require('@ukhomeoffice/taskflow');
-const decorator = require('./decorators');
 const schema = require('@asl/schema');
 const { resubmitted, autoResolved, resolved } = require('./flow/status');
 const tasksRouter = require('./router/tasks');
 const modelTasksRouter = require('./router/model-tasks');
+
+const decorators = require('./decorators');
 
 const profile = require('./middleware/profile');
 
@@ -27,7 +28,8 @@ module.exports = settings => {
 
   const flow = Taskflow({ db: settings.taskflowDB });
 
-  flow.decorate(decorator(settings));
+  flow.decorate(decorators.metadata(settings));
+  flow.decorate(decorators.nextSteps(settings), { list: false });
 
   flow.hook('pre-create', hooks.unique(settings));
