+++ package.json
@@ -26,7 +26,7 @@
   "dependencies": {
     "@asl/schema": "^7.4.3",
     "@asl/service": "^6.17.2",
-    "@ukhomeoffice/taskflow": "^0.10.0",
+    "@ukhomeoffice/taskflow": "^0.11.0",
     "aws-sdk": "^2.270.1",
     "express": "^4.16.4",
     "lodash": "^4.17.11",
